# Pool hash to reduce allocations on Push.