--- a/pkg/ingester/active_series.go
+++ b/pkg/ingester/active_series.go
@@ -1,6 +1,7 @@
 package ingester
 
 import (
+	"hash"
 	"math"
 	"sync"
 	"time"
@@ -60,8 +61,13 @@ func (c *ActiveSeries) UpdateSeries(series labels.Labels, now time.Time, labelsC
 
 var sep = []byte{model.SeparatorByte}
 
+var hashPool = sync.Pool{New: func() interface{} { return xxhash.New() }}
+
 func fingerprint(series labels.Labels) uint64 {
-	sum := xxhash.New()
+	sum := hashPool.Get().(hash.Hash64)
+	defer hashPool.Put(sum)
+
+	sum.Reset()
 	for _, label := range series {
 		_, _ = sum.Write(yoloBuf(label.Name))
 		_, _ = sum.Write(sep)