**What this PR does**: This PR adds pooling to `xxhash.New()`, which helps to reduce allocations on `Push`-path in ingester.

```
name                                  old time/op    new time/op    delta
ActiveSeriesTest_single_series/50-4     86.2ns ± 2%    78.1ns ± 1%    -9.41%  (p=0.000 n=10+8)
ActiveSeriesTest_single_series/100-4    86.5ns ± 1%    80.0ns ±10%    -7.42%  (p=0.002 n=9+10)
ActiveSeriesTest_single_series/500-4    87.3ns ± 6%    90.1ns ±33%      ~     (p=0.393 n=10+10)
ActiveSeries_UpdateSeries-4              736ns ±19%     718ns ±12%      ~     (p=0.953 n=9+10)

name                                  old alloc/op   new alloc/op   delta
ActiveSeriesTest_single_series/50-4      80.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
ActiveSeriesTest_single_series/100-4     80.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
ActiveSeriesTest_single_series/500-4     80.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
ActiveSeries_UpdateSeries-4               241B ±17%      164B ±12%   -31.85%  (p=0.000 n=10+8)

name                                  old allocs/op  new allocs/op  delta
ActiveSeriesTest_single_series/50-4       1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
ActiveSeriesTest_single_series/100-4      1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
ActiveSeriesTest_single_series/500-4      1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
ActiveSeries_UpdateSeries-4               3.00 ± 0%      2.00 ± 0%   -33.33%  (p=0.000 n=10+10)
```

This PR also removes for-loop when updating series timestamp. If there are multiple goroutines trying to update timestamp at the same time, they most likely will be very close to each other... there is no need for super-precision. (And makes benchmarks with many goroutines runnable again)